#147 Refactor Profile Header
